+++ src/app/chatbot-design-studio/chatbot-design-studio.module.ts
@@ -2,7 +2,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { DragDropModule } from '@angular/cdk/drag-drop';
-
+import { A11yModule } from '@angular/cdk/a11y';
 import { RouterModule } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -15,6 +15,7 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatListModule } from '@angular/material/list';
 import { MatRadioModule } from '@angular/material/radio';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
@@ -59,6 +60,9 @@ import { RulesAddComponent } from './cds-rules/rules-add/rules-add.component';
 import { RulesListComponent } from './cds-rules/rules-list/rules-list.component';
 import { ConditionComponent } from './cds-rules/rules-add/condition/condition.component';
 import { ActionComponent } from './cds-rules/rules-add/action/action.component';
+
+
+//intent ACTIONS
 import { ActionWaitComponent } from './cds-dashboard/panel-intent-detail/actions/action-wait/action-wait.component';
 import { ActionEmailComponent } from './cds-dashboard/panel-intent-detail/actions/action-email/action-email.component';
 import { ActionIntentComponent } from './cds-dashboard/panel-intent-detail/actions/action-intent/action-intent.component'
@@ -67,6 +71,9 @@ import { ActionDeleteVariableComponent } from './cds-dashboard/panel-intent-deta
 import { ActionAssignVariableComponent } from './cds-dashboard/panel-intent-detail/actions/action-assign-variable/action-assign-variable.component';
 import { ActionReplaceBotComponent } from './cds-dashboard/panel-intent-detail/actions/action-replace-bot/action-replace-bot.component';
 import { ActionChangeDepartmentComponent } from './cds-dashboard/panel-intent-detail/actions/action-change-department/action-change-department.component';
+import { ActionOnlineAgentsComponent } from './cds-dashboard/panel-intent-detail/actions/action-online-agents/action-online-agents.component';
+import { ActionOpenHoursComponent } from './cds-dashboard/panel-intent-detail/actions/action-open-hours/action-open-hours.component';
+import { ActionHideMessageComponent } from './cds-dashboard/panel-intent-detail/actions/action-hide-message/action-hide-message.component';
 
 @NgModule({
   declarations: [
@@ -75,7 +82,6 @@ import { ActionChangeDepartmentComponent } from './cds-dashboard/panel-intent-de
     ActionReplyComponent,
     TextResponseComponent,
     DelaySliderComponent,
-    // ButtonConfigurationPanelComponent,
     ImageResponseComponent,
     FrameResponseComponent,
     ImageUploadComponent,
@@ -113,9 +119,13 @@ import { ActionChangeDepartmentComponent } from './cds-dashboard/panel-intent-de
     ActionEmailComponent,
     ActionConditionComponent,
     ActionReplaceBotComponent,
-    ActionChangeDepartmentComponent
+    ActionChangeDepartmentComponent,
+    ActionOnlineAgentsComponent,
+    ActionOpenHoursComponent,
+    ActionHideMessageComponent
   ],
   imports: [
+    A11yModule,
     CommonModule,
     DragDropModule,
     NgSelectModule,
@@ -125,6 +135,7 @@ import { ActionChangeDepartmentComponent } from './cds-dashboard/panel-intent-de
     MatSelectModule,
     MatTooltipModule,
     MatRadioModule,
+    MatCheckboxModule,
     MatChipsModule,
     MatGridListModule,
     MatAutocompleteModule,