+++ src/app/models/intent-model.ts
@@ -10,7 +10,6 @@ export class Intent {
     answer?: string;
     form?: Form;
     actions?: Action[];
-    // reply?: Reply;
     id_project?: string;
     language?: string;
     intent_display_name?: string;
@@ -21,38 +20,9 @@ export class Intent {
     id?: string;
     constructor() {
         this.actions = [];
-        //this.reply = new Reply();
     }
-
 }
 
-// export interface iIntent {
-//     webhook_enabled?: boolean;
-//     enabled?: boolean;
-//     topic?: string;
-//     status?: string;
-//     id_faq_kb?: string;
-//     question?: string;
-//     answer?: string;
-//     // form?: Form;
-//     // actions?: Action[];
-//     // reply?: Reply;
-//     id_project?: string;
-//     language?: string;
-//     intent_display_name?: string;
-//     createdBy?: string;
-//     intent_id?: string;
-//     createdAt?: Date;
-//     updatedAt?: Date;
-//     id?: string;
-// }
-
-// export class Question {
-//     split(arg0: string) {
-//       throw new Error('Method not implemented.');
-//     }
-//     type: string;
-// }
 
 export class Action {
     _tdActionType: string;
@@ -90,7 +60,7 @@ export class ActionOnlineAgent extends Action {
     falseIntent: string;
     constructor() {
         super();
-        this._tdActionType = TYPE_ACTION.ONLINE_AGENGS;
+        this._tdActionType = TYPE_ACTION.ONLINE_AGENTS;
     }
 }
 
@@ -100,6 +70,14 @@ export class ActionOpenHours extends Action {
     constructor() {
         super();
         this._tdActionType = TYPE_ACTION.OPEN_HOURS;
+    }
+}
+
+export class ActionHideMessage extends Action {
+  text: string;
+    constructor() {
+        super();
+        this._tdActionType = TYPE_ACTION.HIDE_MESSSAGE;
     }
 }
 
@@ -134,8 +112,10 @@ export class ActionChangeDepartment extends Action {
 }
 
 export class Attributes {
+    disableInputMessage: boolean;
     commands: Command[];
     constructor(commands?: Command[]) {
+        this.disableInputMessage = false;
         this.commands = [];
         if(commands && commands.length>0){
             this.commands = commands;